Modifications

+++ src/main/java/com/cg/go/service/OrderServiceImpl.java
@@ -5,9 +5,8 @@ import java.util.List;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
-import com.cg.go.dao.*;
-import com.cg.go.dao.GrowthReportRepositoryImpl;
-import com.cg.go.dao.IGrowthReportRepository;
+import com.cg.go.dao.IOrderRepository;
+import com.cg.go.dao.OrderRepositoryImpl;
 import com.cg.go.entity.OrderEntity;
 import com.cg.go.exception.OrderException;
 import com.cg.go.util.JpaUtil;
@@ -31,32 +30,81 @@ public class OrderServiceImpl implements IOrderService {
 		return list;
 	}
 
-	public OrderEntity addOrder(OrderEntity orderEntity) throws OrderException{
+	public OrderEntity addOrder(OrderEntity orderEntity){
 		EntityTransaction transaction = entityManager.getTransaction();
-        transaction.begin();
-        OrderEntity order=daoOrder.addOrder(orderEntity);
+		transaction.begin();
+        
+		try {
+	     	   List<OrderEntity> orderlist=daoOrder.findAllOrders();
+	        	if(orderlist.isEmpty()) {
+	        		throw new OrderException("UserId not found");
+	        	}
+	        	OrderEntity order=daoOrder.addOrder(orderEntity);
+	        	return order;
+		 }
+		
+		catch(OrderException orderException) {
+            System.out.println(orderException.getMessage());
+            transaction.commit();
+    	} 
         transaction.commit();
-		return order;
+		return new OrderEntity();
 	}
 
-	public void deleteAllOrders() throws OrderException{
+	public void deleteAllOrders(){
 		EntityTransaction transaction = entityManager.getTransaction();
         transaction.begin();
-        daoOrder.deleteAllOrders();
-        transaction.commit();
+        try {
+     	   List<OrderEntity> orderlist=daoOrder.findAllOrders();
+        	if(orderlist.isEmpty()) {
+        		throw new OrderException("No Order in the list to delete");
+        	}
+            daoOrder.deleteAllOrders();	
+            transaction.commit();
+
+        	}
+        	catch(OrderException orderException) {
+                System.out.println(orderException.getMessage());
+                transaction.commit();
+        	}
+        
+        	transaction.commit();
+
 	}
 
-	public void deleteOrderById(String orderId) throws OrderException{
+	public void deleteOrderById(String orderId) {
 		EntityTransaction transaction = entityManager.getTransaction();
         transaction.begin();
-        daoOrder.deleteOrderById(orderId);
-        transaction.commit();
+       try {
+    	   OrderEntity orderEntity=entityManager.find(OrderEntity.class,orderId);
+	       	if(orderEntity==null) {
+	       		throw new OrderException("OrderId not found");
+	       	}
+    	   daoOrder.deleteOrderById(orderId);
+       }
+       catch(OrderException orderException) {
+    	   System.out.println(orderException.getMessage());
+    	   transaction.commit();
+       }
+    	   transaction.commit();
 	}
 
-	public void updateDate(String orderId, LocalDate dispatchDate, LocalDate arrivalDate) throws OrderException{
+	
+	
+	public void updateDate(String orderId, LocalDate dispatchDate, LocalDate arrivalDate) {
 		EntityTransaction transaction = entityManager.getTransaction();
         transaction.begin();
-        daoOrder.updateDate(orderId, dispatchDate, arrivalDate);
+        try {
+     	   OrderEntity orderEntity=entityManager.find(OrderEntity.class,orderId);
+ 	       	if(orderEntity==null) {
+ 	       		throw new OrderException("OrderId not found");
+ 	       	}
+ 	       	daoOrder.updateDate(orderId, dispatchDate, arrivalDate);
+        }
+        catch(OrderException orderException) {
+        	System.out.println(orderException.getMessage());
+    	   transaction.commit();
+        }
         transaction.commit();
 	}
